Massive refactoring to allow for benchmarks
This commit refactors the code to make the main package more of a
generic fractal "viewer". This abstracts out all the fractal specific
bits, so the viewer just accepts a generic fractal interface, but
doesn't need to know the specific implementation details of a given
fractal. The user can specify which fractal they'd like to view
(currently this is just the Mandelbrot Set). As part of this I've added
the fractal package, which contains the fractal interface, and then
different fractals will implement this interface. This also makes it
easier to test/benchmark the Mandelbrot "draw" function.